+++ src/main/java/com/ironSync/controller/WorkoutController.java
@@ -1,7 +1,9 @@
 package com.ironSync.controller;
 
+import com.ironSync.dto.WorkoutDTO;
 import com.ironSync.model.Workout;
 import com.ironSync.model.WorkoutEntry;
+import com.ironSync.storage.FileManager;
 
 import java.util.List;
 
@@ -9,6 +11,8 @@ import java.util.List;
  * Controller responsible for creating Workout objects.
  */
 public class WorkoutController {
+
+    private final FileManager fileManager = new FileManager();
 
     /**
      * Builds a new {@link Workout} instance from the provided parameters.
@@ -20,6 +24,14 @@ public class WorkoutController {
      * @return A new Workout object constructed with the provided data.
      */
     public Workout workoutBuilder(List<WorkoutEntry> workoutEntries, String workoutTitle, String workoutNotes, String workoutDate) {
+        Workout createdWorkout = new Workout(workoutEntries, workoutTitle, workoutNotes, workoutDate);
+        WorkoutDTO workoutDTO = new WorkoutDTO(createdWorkout);
+        registerWorkout(workoutDTO);
+
         return new Workout(workoutEntries, workoutTitle, workoutNotes, workoutDate);
+    }
+
+    public void registerWorkout(WorkoutDTO workoutDTO) {
+        fileManager.registerSerializedWorkout(workoutDTO);
     }
 }