+++ src/main/java/com/ironSync/storage/JsonSerializer.java
@@ -1,0 +1,20 @@
+package com.ironSync.storage;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public class JsonSerializer {
+    private final ObjectMapper mapper;
+
+    public JsonSerializer() {
+        this.mapper = new ObjectMapper();
+    }
+
+    public String serialize(Object serializableObject) {
+        try {
+            return this.mapper.writeValueAsString(serializableObject);
+        } catch (JsonProcessingException e) {
+            throw new RuntimeException("An error happened when serializing the object to JSON.");
+        }
+    }
+}